+++ hoteli-service/src/main/java/com/isa/hoteli/hoteliservice/controller/TipSobeController.java
@@ -3,6 +3,8 @@ package com.isa.hoteli.hoteliservice.controller;
 import java.util.ArrayList;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +15,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.isa.hoteli.hoteliservice.avio.model.Korisnik;
+import com.isa.hoteli.hoteliservice.avio.model.Rola;
+import com.isa.hoteli.hoteliservice.avio.service.KorisnikService;
 import com.isa.hoteli.hoteliservice.dto.TipSobeDTO;
 import com.isa.hoteli.hoteliservice.model.TipSobe;
 import com.isa.hoteli.hoteliservice.service.TipSobeService;
@@ -24,6 +29,9 @@ public class TipSobeController {
 
 	@Autowired
 	private TipSobeService tipSobeService;
+	
+	@Autowired
+	private KorisnikService korisnikService;
 	
 	@RequestMapping(value="/all", method = RequestMethod.GET)
 	public ResponseEntity<List<TipSobeDTO>> getTypes(){
@@ -56,19 +64,27 @@ public class TipSobeController {
 	}
 	
 	@RequestMapping(value="/", method = RequestMethod.POST)
-	public ResponseEntity<TipSobeDTO> createType(@RequestBody TipSobeDTO dto){
-		TipSobe tipSobe = new TipSobe(dto);
-		TipSobeDTO returnType = tipSobeService.createType(tipSobe);
-		if(returnType!=null) {
-			return new ResponseEntity<>(returnType, HttpStatus.OK);
+	public ResponseEntity<TipSobeDTO> createType(@RequestBody TipSobeDTO dto, HttpServletRequest req){
+		Korisnik k = korisnikService.zaTokene(req);
+		if(k!=null && k.getRola().equals(Rola.ADMIN_HOTELA) && k.getZaduzenZaId()==dto.getHotel().getId()) {
+			TipSobe tipSobe = new TipSobe(dto);
+			TipSobeDTO returnType = tipSobeService.createType(tipSobe);
+			if(returnType!=null) {
+				return new ResponseEntity<>(returnType, HttpStatus.OK);
+			}
+			
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
 		}
-		
-		return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@RequestMapping(value="/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<String> deleteTypeById(@PathVariable("id") Long id){
-		return new ResponseEntity<String>(tipSobeService.deleteType(id), HttpStatus.OK);
+	public ResponseEntity<String> deleteTypeById(@PathVariable("id") Long id, HttpServletRequest req){
+		Korisnik k = korisnikService.zaTokene(req);
+		if(k!=null && k.getRola().equals(Rola.ADMIN_HOTELA)) {
+			return new ResponseEntity<String>(tipSobeService.deleteType(id), HttpStatus.OK);
+		}
+		return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
 	}
 	
 	@RequestMapping(value="/{id}", method = RequestMethod.PUT)
